+++ src/runtime/values/basevalues/value.py
@@ -245,9 +245,10 @@ class Value:
         """
         return None, self.illegal_operation()
 
-    def execute(self, args: list[Value], interpreter_: type[Interpreter], run: RunFunction, noug_dir: str,
-                lexer_metas: dict[str, str | bool], exec_from: str = "<invalid>", use_context: Context | None = None,
-                cli_args: list[String] | None = None, work_dir: str | None = None):
+    def execute(self, args: list[Value | tuple[String, Value]], interpreter_: type[Interpreter], run: RunFunction,
+                noug_dir: str, lexer_metas: dict[str, str | bool], exec_from: str = "<invalid>",
+                use_context: Context | None = None, cli_args: list[String] | None = None,
+                work_dir: str | None = None):
         """Execute the function.
         Returns a result"""
         return RTResult().failure(self.illegal_operation())