+++ src/runtime/values/functions/base_builtin_func.py
@@ -42,9 +42,10 @@ class BaseBuiltInFunction(BaseFunction):
     def get_comparison_ne(self, other: Value):
         return Number(not self.is_eq(other), self.pos_start, other.pos_end).set_context(self.context), None
 
-    def execute(self, args: list[Value], interpreter_: type[Interpreter], run: RunFunction, noug_dir: str,
-                lexer_metas: dict[str, str | bool], exec_from: str = "<invalid>", use_context: Context | None = None,
-                cli_args: list[String] | None = None, work_dir: str | None = None):
+    def execute(self, args: list[Value | tuple[String, Value]], interpreter_: type[Interpreter], run: RunFunction,
+                noug_dir: str, lexer_metas: dict[str, str | bool], exec_from: str = "<invalid>",
+                use_context: Context | None = None, cli_args: list[String] | None = None,
+                work_dir: str | None = None):
         return RTResult().success(NoneValue(self.pos_start, self.pos_end, False))
 
     def no_visit_method(self, exec_ctx: Context):